requirements.txt
PyMuPDF==1.23.21
pdf2image==1.17.0
Pillow==10.2.0
transformers==4.41.2
sentence-transformers==2.7.0
torch==2.3.0
streamlit==1.35.0
python-dotenv==1.0.1
openai==1.30.1
requests>=2.31.0
